app/src/main/java/com/cshawn/leetcodes/sword/Sword_28.java
package com.cshawn.leetcodes.sword;

/**
 * 对称的二叉树
 * 请实现一个函数，用来判断一棵二叉树是不是对称的。如果一棵二叉树和它的镜像一样，那么它是对称的。
 *
 * 例如，二叉树 [1,2,2,3,4,4,3] 是对称的。
 *
 *     1
 *    / \
 *   2   2
 *  / \ / \
 * 3  4 4  3
 * 但是下面这个 [1,2,2,null,3,null,3] 则不是镜像对称的:
 *
 *     1
 *    / \
 *   2   2
 *    \   \
 *    3    3
 *
 *  
 *
 * 示例 1：
 *
 * 输入：root = [1,2,2,3,4,4,3]
 * 输出：true
 * 示例 2：
 *
 * 输入：root = [1,2,2,null,3,null,3]
 * 输出：false
 *  
 *
 * 限制：
 *
 * 0 <= 节点个数 <= 1000
 *
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/dui-cheng-de-er-cha-shu-lcof
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 * @author C.Shawn
 * @Date 2020/12/10 22:30
 */
public class Sword_28 {
    public boolean isSymmetric(TreeNode root) {
        if (root == null) {
            return false;
        }
        return isSymmetric(root.left, root.right);
    }

    /**
     * 判断两个对称位置上的两个结点是否对称
     */
    private boolean isSymmetric(TreeNode node1, TreeNode node2) {
        if (node1 == null) {
            return node2 == null;
        }
        // 当前结点值相等
        return node2 != null && node1.val == node2.val
                // 且1的左子树与2的右子树对称
                && isSymmetric(node1.left, node2.right)
                // 且1的右子树与2的左子树对称
                && isSymmetric(node1.right, node2.left);
    }
}
